Fixes grayscale calculation
Until now, the calculation of the gray value was different between
debug and release builds: the latter used to be more aggressive in the
conversion, resulting in lighter whites and darker blacks.

This issue seems to have been cause by the optimization profile of
the release build type. To fix this, the calculation has been split into more
detailed steps so that optimizations don't make too much rounding
errors. Although I admit that why this refactoring is effective is not yet
clear to me.

Also, while researching for the above mentioned problem, I found that
the Rec. 709 formula with primary coefficients seems to create a more
accurate grayscale representation. This new formula with specific
coefficients was also applied with this commit.