Refactors MainViewModel
To reflect the new CommandBar of the MainPage,
Change the SaveButtonVisibility to SaveButtonIsEnabled. This is because we don't want to hide
anymore the button to the user. Instead, we want to show
what actions can be performed once other actions have been
triggered.

For this same reason, the ConvertButtonVisibility has been
removed entirely.